requirements.txt
streamlit
numpy
pandas
faker
folium
streamlit_foliu